+++ CHANGELOG.md
@@ -43,6 +43,7 @@ CHANGELOG
 - Travis: Use `intelmqsetup` here too.
   - Install required build dependencies for the Debian package build test.
 - Added tests for the new bot `intelmq.bots.outputs.misp.output_feed` (#1473).
+- Added tests for the new bot `intelmq.bots.experts.misp.expert` (#1473).
 
 ### Tools
 - `intelmqctl`: